requirements.txt
# Core
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.41.0
python-multipart>=0.0.17,<0.1.0
pydantic-settings>=2.6.0,<3.0.0

# Database
sqlalchemy[asyncio]>=2.0.36,<2.1.0
aiosqlite>=0.20.0,<0.21.0

# Templates
jinja2>=3.1.4,<3.2.0

# Images
pillow>=11.0.0,<12.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Utils
python-slugify>=8.0.4,<9.0.0
python-dotenv>=1.0.1,<2.0.0

# Export/Import
openpyxl>=3.1.5,<4.0.0
